+++ src/Assert/Assert.ts
@@ -1,41 +1,43 @@
 import {MessageProcessor} from "./index";
 import {InvalidArgumentException} from "../Exception/index";
+import {ObjectComparator} from "../Util/ObjectComparator";
 
 export class Assert {
 
     public static assertTrue(expr: boolean, msg: string): void {
         if (expr !== true) {
-            let description = "Failed asserting that value is true. ";
+            let description = "Failed asserting that '" + expr + "' is true. ";
             MessageProcessor.handleFailedAssertion(msg, description);
         }
     }
 
     public static assertFalse(expr: boolean, msg: string): void {
         if (expr !== false) {
-            let description = "Failed asserting that value is false. ";
+            let description = "Failed asserting that '" + expr + "' is false. ";
             MessageProcessor.handleFailedAssertion(msg, description);
         }
     }
 
     public static assertUndefined(expr, msg: string): void {
         if (expr !== undefined) {
-            let description = "Failed asserting that value is undefined. ";
+            let description = "Failed asserting that '" + expr + "' is undefined. ";
             MessageProcessor.handleFailedAssertion(msg, description);
         }
     }
 
     public static assertEquals(expected, actual, msg: string): void {
+        // TODO extract diff building to special class with separate unit tests.
         let descriptionLines = [
             "Failed asserting that two objects are equals. ",
             "--- Expected",
             "+++ Actual",
             "@@ @@",
-            "- " + expected,
-            "+ " + actual
+            "-" + JSON.stringify(expected, null, 2),
+            "+" + JSON.stringify(actual, null, 2),
         ];
         let description = descriptionLines.join("\n");
         if (expected instanceof Object && actual instanceof Object) {
-            if (Comparator.isEquals(expected, actual) === false) {
+            if (ObjectComparator.isEquals(expected, actual) === false) {
                 MessageProcessor.handleFailedAssertion(msg, description);
             }
         } else if (expected != actual) {
@@ -66,7 +68,7 @@ export class Assert {
 
     public static assertNull(expr, msg: string): void {
         if (expr !== null) {
-            let description = "Failed asserting that value is null. ";
+            let description = "Failed asserting that '" + expr + "' is null. ";
             MessageProcessor.handleFailedAssertion(msg, description);
         }
     }
@@ -90,37 +92,5 @@ export class Assert {
             let description = "Failed asserting that object instance of " + expected.prototype.constructor.name + ". ";
             MessageProcessor.handleFailedAssertion(msg, description);
         }
-    }
-}
-
-/**
- * Thanks to Joshua Clanton and his Drips:
- * this objects comparison based on code from http://adripofjavascript.com/blog/drips/object-equality-in-javascript.html
- */
-class Comparator {
-    public static isEquals(expected, actual): boolean {
-        // Create arrays of property names
-        let expectedProps = Object.getOwnPropertyNames(expected);
-        let actualProps = Object.getOwnPropertyNames(actual);
-
-        // If number of properties is different,
-        // objects are not equivalent
-        if (expectedProps.length != actualProps.length) {
-            return false;
-        }
-
-        for (let i = 0; i < expectedProps.length; i++) {
-            let propName = expectedProps[i];
-
-            // If values of same property are not equal,
-            // objects are not equivalent
-            if (expected[propName] !== actual[propName]) {
-                return false;
-            }
-        }
-
-        // If we made it this far, objects
-        // are considered equivalent
-        return true;
     }
 }