+++ src/Assert/DomAssert.ts
@@ -4,13 +4,13 @@ export class DomAssert {
 
     public static assertElementExists(context: NodeSelector, selector: string, msg: string): void {
         if (this.getElementsCount(context, selector) === 0) {
-            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that element with selector " + selector + " exists in the document. ");
+            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that element with selector '" + selector + "' exists in the document. ");
         }
     }
 
     public static assertElementNotExists(context: NodeSelector, selector: string, msg: string): void {
         if (this.getElementsCount(context, selector) > 0) {
-            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that element with selector " + selector + " not exists in the document. ");
+            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that element with selector '" + selector + "' not exists in the document. ");
         }
     }
 
@@ -22,13 +22,13 @@ export class DomAssert {
 
     public static assertElementNotHasClass(element: Element, className: string, msg: string): void {
         if (element.classList.contains(className)) {
-            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that element not has this class '" + className + "'. ");
+            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that element hasn't class '" + className + "'. ");
         }
     }
 
     public static assertElementsCount(document: NodeSelector, selector: string, count: number, msg: string): void {
         if (this.getElementsCount(document, selector) !== count) {
-            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that number of elements (" + selector + ") is " + count + ". ");
+            MessageProcessor.handleFailedAssertion(msg, "Failed asserting that number of elements '" + selector + "' is " + count + ". ");
         }
     }