+++ tests/Assert/AssertTest.ts
@@ -1,5 +1,6 @@
 import {AbstractUnitTestCase} from "../../dist/index";
 import {FailedAssertionException} from "../../dist/Assert/index";
+import {InvalidArgumentException} from "../../dist/Exception/InvalidArgumentException";
 
 export class AssertTest extends AbstractUnitTestCase {
 
@@ -15,7 +16,8 @@ export class AssertTest extends AbstractUnitTestCase {
             expectedException = e;
         } finally {
             this.assertInstanceOf(FailedAssertionException, expectedException);
-            this.assertEquals("oops!\nFailed asserting that value is true. ", expectedException.getMessage());
+            let message = expectedException.getMessage();
+            this.assertEquals("oops!\nFailed asserting that 'false' is true. ", message);
         }
     }
 
@@ -23,8 +25,34 @@ export class AssertTest extends AbstractUnitTestCase {
         this.assertFalse(false);
     }
 
+    public testAssertFalseFail() {
+        let expectedException;
+        try {
+            this.assertFalse(true, "oops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertEquals("oops!\nFailed asserting that 'true' is false. ", message);
+        }
+    }
+
     public testAssertUndefined() {
         this.assertUndefined(this[100500]);
+    }
+
+    public testAssertUndefinedFail() {
+        let expectedException;
+        try {
+            this.assertUndefined(true, "oops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertEquals("oops!\nFailed asserting that 'true' is undefined. ", message);
+        }
     }
 
     public testAssertEquals() {
@@ -36,10 +64,47 @@ export class AssertTest extends AbstractUnitTestCase {
         };
 
         this.assertEquals(expected, actual);
-
         this.assertEquals("foo", "foo");
-
         this.assertEquals(1, "1");
+    }
+
+    public testAssertEqualsFailWithObjectsDiff() {
+        let expected = {
+            foo: "bar"
+        };
+        let actual = {
+            foo: "not bar"
+        };
+        let expectedException;
+
+        try {
+            this.assertEquals(expected, actual, "oops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertContains("oops!\nFailed asserting that two objects are equals. ", message);
+            this.assertContains("+{\n  \"foo\": \"not bar\"\n}", message);
+        }
+    }
+
+    public testAssertEqualsFailWithScalarDiff() {
+        let expected = 1;
+        let actual = 2;
+        let expectedException;
+
+        try {
+            this.assertEquals(expected, actual, "oops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertContains("oops!\nFailed asserting that two objects are equals. ", message);
+            this.assertContains("-1", message);
+            this.assertContains("+2", message);
+        }
     }
 
     public testAssertSame() {
@@ -47,16 +112,73 @@ export class AssertTest extends AbstractUnitTestCase {
         this.assertSame(baz, baz);
     }
 
+    public testAssertSameFail() {
+        let expected = 1;
+        let actual = 2;
+        let expectedException;
+
+        try {
+            this.assertSame(expected, actual, "oops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertContains("oops!", message);
+            this.assertContains("Failed asserting that two objects are same. ", message);
+        }
+    }
+
     public testAssertNotSame() {
         this.assertNotSame("foo", "bar");
+    }
+
+    public testAssertNotSameFail() {
+        let expectedException;
+
+        try {
+            this.assertNotSame(1, 1, "oops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertContains("oops!\nFailed asserting that objects are not same. ", message);
+        }
     }
 
     public testAssertNotNull() {
         this.assertNotNull("foo");
     }
 
+    public testAssertNotNullFail() {
+        let expectedException;
+        try {
+            this.assertNotNull(null, "ooops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertContains("ooops!\nFailed asserting that value is not null.", message);
+        }
+    }
+
     public testAssertNull() {
         this.assertNull(null);
+    }
+
+    public testAssertNullFail() {
+        let expectedException;
+        try {
+            this.assertNull(true, "ooops!");
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+            let message = expectedException.getMessage();
+            this.assertContains("ooops!\nFailed asserting that 'true' is null.", message);
+        }
     }
 
     public testAssertContainsNeedleInStringObject() {
@@ -71,6 +193,28 @@ export class AssertTest extends AbstractUnitTestCase {
         this.assertContains("foo", ["bar", "foo"]);
     }
 
+    public testAssertContainsFail() {
+        let expectedException;
+        try {
+            this.assertContains(2, [0, 1]);
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(FailedAssertionException, expectedException);
+        }
+    }
+
+    public testAssertContainsInvalidArgument() {
+        let expectedException;
+        try {
+            this.assertContains("needle", 100500);
+        } catch (e) {
+            expectedException = e;
+        } finally {
+            this.assertInstanceOf(InvalidArgumentException, expectedException);
+        }
+    }
+
     public testAssertInstanceOf() {
         let foo = new FooImpl();
         this.assertInstanceOf(AbstractFoo, foo);
@@ -83,10 +227,11 @@ export class AssertTest extends AbstractUnitTestCase {
         try {
             this.assertInstanceOf(AbstractUnitTestCase, foo);
         } catch (e) {
-            expectedException = <FailedAssertionException> e;
-            this.assertEquals(e.getMessage(), "Failed asserting that object instance of AbstractUnitTestCase. ");
+            expectedException = e;
         } finally {
             this.assertInstanceOf(FailedAssertionException, expectedException, "FailedAssertionException wasn't throw on wrong assertion. ");
+            let message = expectedException.getMessage();
+            this.assertEquals("Failed asserting that object instance of AbstractUnitTestCase. ", message);
         }
     }